+++ lib/ui/component/base_drawer.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:smart_order_app/ui/component/error.dart';
 import 'package:smart_order_app/ui/component/loader.dart';
-import 'package:smart_order_app/ui/page/order_support_select/page.dart';
-import 'package:smart_order_app/ui/page/phrase/add/page.dart';
-import 'package:smart_order_app/ui/page/scene/add/page.dart';
+import 'package:smart_order_app/ui/page/order/select/page.dart';
+import 'package:smart_order_app/ui/page/management/phrase/add/page.dart';
+import 'package:smart_order_app/ui/page/management/scene/add/page.dart';
 import 'package:smart_order_app/usecase/state/scenes.dart';
 
 class BaseDrawer extends ConsumerWidget {
@@ -18,8 +18,8 @@ class BaseDrawer extends ConsumerWidget {
         children: [
           const ListTile(title: Text("◼️場面選択")),
           scenesFuture.when(
-            error: (e, s) => createErrorMessage(),
-            loading: () => createProgressIndicator(context),
+            error: (e, s) => buildErrorMessage(),
+            loading: () => buildProgressIndicator(context),
             data: (scenes) {
               return Column(
                 children: scenes.map((scene) {
@@ -30,7 +30,7 @@ class BaseDrawer extends ConsumerWidget {
                         context,
                         MaterialPageRoute(
                             builder: (context) =>
-                                OrderSupportSelectPage(sceneName: scene.scene)),
+                                OrderSelectPage(sceneName: scene.scene)),
                       );
                     },
                   );
